requirements.txt
cryptography==36.0.0
Django==3.2.9
PySocks==1.7.1
pyTelegramBotAPI==4.2.0
python-telegram-bot==13.8.1
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.7
python-dotenv
requests
